app/backend/requirements.txt
Django~=4.1
psycopg2>=2.8
django-cities-light